Restructure INARA panels under panels namespace
