+++ FEATURES.md
@@ -51,7 +51,7 @@ Use these shortnames when coordinating INARA work:
 - **CARGO_LEDGER – Cargo Hold Valuation.** `CargoHoldPanel` pulls the live ship loadout and cargo inventory, derives a memoized cargo fingerprint, and requests `/api/glitch-commodity-values` to merge INARA submissions with in-game journal market logs. The valuation response contains INARA and local market health indicators; present both statuses in the UI so commanders can reconcile stale remote intel. Cache-heavy helpers (`isSameMarketEntry`, `mergeInventoryRows`) ensure we do not thrash the DOM when only metadata shifts. Keep the utilisation meter at the top of the panel in sync with the ship's capacity so miners can instantly judge how much space remains.
 - **MISSION_BEACON – Mining Mission Radar.** `MissionsPanel` watches the current system via `useSystemSelector`, hydrates faction reputation via `/api/faction-standings`, and caches the last eight system lookups in `localStorage`. INARA fetches stream in via `/api/glitch-missions` POST requests, automatically downgrading to cached payloads on errors. Maintain the status machine (`idle`, `loading`, `empty`, `error`, `populated`) so accessibility strings stay accurate.
 - **PRISTINE_TRACKER – Ring Prospecting.** `PristineMiningPanel` (lower in `inara.js`) cross-references INARA pristine mining listings with ICARUS system-map intel (`SystemMapProvider`). Rows expand into detail drawers populated via `NavigationInspectorPanel`, so ensure new fields are wired through that provider rather than injecting ad-hoc fetches. Keep an eye on `animateTableEffect()` hooks to preserve the neon scan reveal.
-- **RADIO_RELAY – Pirate Radio Broadcast Deck.** `PirateRadioPanel` (`src/client/components/glitch/pirate-radio.js`) hydrates the underground audio stream through `sendEvent('getPirateRadioPlaylist')`, wires directory management to `setPirateRadioDirectories`, and triggers rescans with `rescanPirateRadio`. It listens for `pirateRadioUpdate`/`pirateRadioDirectoriesUpdated` broadcasts to stay in sync and leans on the INARA panel shell styles so the navigation rail spacing remains consistent. Preserve the autoplay-on-ended flow when extending the deck so broadcasts continue looping seamlessly.
+- **RADIO_RELAY – Pirate Radio Broadcast Deck.** `PirateRadioPanel` (`src/client/components/panels/inara/pirate-radio.js`) hydrates the underground audio stream through `sendEvent('getPirateRadioPlaylist')`, wires directory management to `setPirateRadioDirectories`, and triggers rescans with `rescanPirateRadio`. It listens for `pirateRadioUpdate`/`pirateRadioDirectoriesUpdated` broadcasts to stay in sync and leans on the INARA panel shell styles so the navigation rail spacing remains consistent. Preserve the autoplay-on-ended flow when extending the deck so broadcasts continue looping seamlessly.
 - **UPLINK_FEED – Ambient Telemetry Overlay.** The uplink console (`inara.js` final sections) rotates pseudo-telemetry headlines, user-configurable cadence controls, and integrates with `glitchTickerMessages`. Additions should honor the animation timings and respect the reduced-motion guard.
 - **ASSIMILATION_GATE – Page Shell & Arrival Sequence.** `GlitchPage` toggles the global theme class, triggers arrival animations, and manages top-level tab state. Extend it via composition—drop new sections into the existing `<Panel>` layout so navigation/ARIA wiring continues to work.
 - **TAB_SHELL – Tab Navigation.** The `glitchTabs` array describes the tab structure and icons; updates must keep the keyboard handlers (`handleTabKeyPress`) intact. When adding tabs, double-check breakpoints so the secondary nav remains scrollable on narrow widths.
