+++ src/client/components/panels/inara/item-cell.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { sanitizeInaraText } from '../../lib/sanitize-inara-text'
+import { sanitizeInaraText } from 'lib/sanitize-inara-text'
 import StackedCell from './stacked-cell'
 import styles from './trade-route-cells.module.css'
