+++ src/client/components/panels/inara/stacked-cell.js
