+++ src/client/components/panels/inara/station-summary.module.css
