+++ src/client/components/panels/inara/transfer-context-summary.module.css
