+++ src/client/pages/inara.js
@@ -3,11 +3,11 @@ import Layout from '../components/layout'
 import Panel from '../components/panel'
 import PanelNavigation from '../components/panel-navigation'
 import Icons from '../lib/icons'
-import TransferContextSummary from '../components/glitch/transfer-context-summary'
-import StationSummary, { StationIcon, DemandIndicator } from '../components/glitch/station-summary'
-import CommoditySummary, { CommodityIcon } from '../components/glitch/commodity-summary'
+import TransferContextSummary from '../components/panels/inara/transfer-context-summary'
+import StationSummary, { StationIcon, DemandIndicator } from '../components/panels/inara/station-summary'
+import CommoditySummary, { CommodityIcon } from '../components/panels/inara/commodity-summary'
 import CopyOnClick from '../components/copy-on-click'
-import PirateRadioPanel from '../components/glitch/pirate-radio'
+import PirateRadioPanel from '../components/panels/inara/pirate-radio'
 import NavigationInspectorPanel from '../components/panels/nav/navigation-inspector-panel'
 import animateTableEffect from '../lib/animate-table-effect'
 import { useSocket, sendEvent, eventListener } from '../lib/socket'
